+++ website/docusaurus.config.js
@@ -119,6 +119,8 @@ module.exports = {
           { from: ['/docs', '/docs/about', '/docs/basics'], to: '/docs/basics/about'},
           { from: '/docs/language', to: '/docs/language/contracts' },
           { from: '/docs/sdk', to: '/docs/sdk/instantiation' },
+          { from: '/docs/sdk/transactions', to: '/docs/releases/migration-notes' },
+          { from: '/docs/sdk/transactions-advanced', to: '/docs/sdk/transaction-builder' },
           { from: '/docs/guides', to: '/docs/guides/covenants' },
           { from: '/docs/getting-started', to: '/docs/basics/getting-started' },
           { from: '/docs/examples', to: '/docs/language/examples' },