+++ website/docs/releases/release-notes.md
@@ -12,6 +12,7 @@ This update adds CashScript support for the new BCH 2025 network upgrade. To rea
 #### CashScript SDK
 - :hammer_and_wrench: Update debug tooling to work with the new `BCH_2025_05` instruction set.
 - :boom: **BREAKING**: Remove support for old contracts compiled with CashScript v0.6.x or earlier.
+- :hammer_and_wrench: Deprecate the simple transaction builder. You can still use the simple transaction builder with the current SDK, but this support will be removed in a future release.
 
 ## v0.10.5
 
@@ -108,7 +109,7 @@ https://x.com/CashScriptBCH/status/1833454128426615174
 ## v0.9.0
 
 #### CashScript SDK
-- :sparkles: Add new advanced `TransactionBuilder` class that allows combining UTXOs from multiple different smart contracts and P2PKH UTXOs in a single transaction.
+- :sparkles: Add new `TransactionBuilder` class that allows combining UTXOs from multiple different smart contracts and P2PKH UTXOs in a single transaction.
 - :hammer_and_wrench: Deprecate all `meep` functionality. Meep has been unmaintained for years and does not support many new CashScript features. Meep functionality will be removed in a future release.
 
 ---